Print transcript instead of logging for transcribe CLI  (#156)
### Why is this change needed?
<!-- Explain the motivation for this change. What problem does it solve?
-->

When building with -c release we are intentionally not logging in CLI,
this is problematic as it doesn't show the final results when running in
the CLI

https://github.com/FluidInference/FluidAudio/issues/154

```
swift run -c release fluidaudio transcribe yc_first_minute.wav --output foo.json | tee bar.json
[1/1] Planning build
Building for production...
[5/5] Linking fluidaudio
Build of product 'fluidaudio' complete! (9.44s)
In the last seven days, we've signed the same number of contracts as we signed in the Hall of Q4. There is clear tangible value being driven by these products and it's only gonna get better and quickly. Ultimately you've got to be very passionate have that perseverance. And if that sounds good to you, then then build a startup. If something logically makes sense, you should probably continue doing that thing, right? And not let anything stop you. And I think like the consistency that we've noticed of founders that we've some of that we've invested in or work with is like the ones that kind of do that and really persevere tend to win. Today we're here with Arnie and Chas Englander. Uh they are the founders of Model ML from Winter 24. Um, they started two other YC companies that both were successful and sold, Fancy and Fat Lama. And this is probably the first time I've worked with a company where both of the founders had had a previous successful uh YC company before. So I'm super excited to.

brandonweng@Brandons-MacBook-Pro FluidAudio % cat bar.json
In the last seven days, we've signed the same number of contracts as we signed in the Hall of Q4. There is clear tangible value being driven by these products and it's only gonna get better and quickly. Ultimately you've got to be very passionate have that perseverance. And if that sounds good to you, then then build a startup. If something logically makes sense, you should probably continue doing that thing, right? And not let anything stop you. And I think like the consistency that we've noticed of founders that we've some of that we've invested in or work with is like the ones that kind of do that and really persevere tend to win. Today we're here with Arnie and Chas Englander. Uh they are the founders of Model ML from Winter 24. Um, they started two other YC companies that both were successful and sold, Fancy and Fat Lama. And this is probably the first time I've worked with a company where both of the founders had had a previous successful uh YC company before. So I'm super excited to.
```